Rollback refactor
