+++ app/src/main/java/com/maximo/douglas/githubconsumer/di/ViewModelModule.kt
@@ -1,19 +1,19 @@
 package com.maximo.douglas.githubconsumer.di
 
-import com.maximo.douglas.features.gitpullrequests.SelectGitPullRequestViewModel
-import com.maximo.douglas.features.gitrepositories.SelectGitRepositoryViewModel
+import com.maximo.douglas.githubconsumer.ui.gitpullrequests.SelectGitPullRequestViewModel
+import com.maximo.douglas.githubconsumer.ui.gitrepositories.SelectGitRepositoryViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
 
 val viewModelModule = module {
     viewModel {
-        com.maximo.douglas.features.gitrepositories.SelectGitRepositoryViewModel(
+        SelectGitRepositoryViewModel(
             get(),
             get()
         )
     }
     viewModel {
-        com.maximo.douglas.features.gitpullrequests.SelectGitPullRequestViewModel(
+        SelectGitPullRequestViewModel(
             get(),
             get()
         )
