+++ app/src/main/java/com/maximo/douglas/githubconsumer/ui/gitpullrequests/adapters/SelectGitPullRequestAdapter.kt
@@ -1,4 +1,4 @@
-package com.maximo.douglas.features.gitpullrequests.adapters
+package com.maximo.douglas.githubconsumer.ui.gitpullrequests.adapters
 
 import android.view.LayoutInflater
 import android.view.View
@@ -6,10 +6,10 @@ import android.view.ViewGroup
 import androidx.databinding.DataBindingUtil
 import androidx.recyclerview.widget.RecyclerView
 import androidx.recyclerview.widget.RecyclerView.Adapter
+import br.com.argmax.githubconsumer.R
+import br.com.argmax.githubconsumer.databinding.GitPullRequestCardViewHolderBinding
 import com.maximo.douglas.domain.entity.gitpullrequest.GitPullRequest
-import com.maximo.douglas.features.R
-import com.maximo.douglas.features.databinding.GitPullRequestCardViewHolderBinding
-import com.maximo.douglas.features.gitpullrequests.listeners.OnPullRequestClickListener
+import com.maximo.douglas.githubconsumer.ui.gitpullrequests.listeners.OnPullRequestClickListener
 import kotlinx.android.extensions.LayoutContainer
 import kotlinx.android.synthetic.main.git_pull_request_card_view_holder.*
