+++ app/src/main/java/com/maximo/douglas/githubconsumer/ui/gitrepositories/adapters/SelectGitRepositoryAdapter.kt
@@ -1,4 +1,4 @@
-package com.maximo.douglas.features.gitrepositories.adapters
+package com.maximo.douglas.githubconsumer.ui.gitrepositories.adapters
 
 import android.view.LayoutInflater
 import android.view.View
@@ -6,10 +6,10 @@ import android.view.ViewGroup
 import androidx.databinding.DataBindingUtil
 import androidx.recyclerview.widget.RecyclerView
 import androidx.recyclerview.widget.RecyclerView.Adapter
+import br.com.argmax.githubconsumer.R
+import br.com.argmax.githubconsumer.databinding.GitRepositoryCardViewHolderBinding
 import com.maximo.douglas.domain.entity.gitrepository.GitRepository
-import com.maximo.douglas.features.R
-import com.maximo.douglas.features.databinding.GitRepositoryCardViewHolderBinding
-import com.maximo.douglas.features.gitrepositories.listeners.OnGitRepositoryClickListener
+import com.maximo.douglas.githubconsumer.ui.gitrepositories.listeners.OnGitRepositoryClickListener
 import kotlinx.android.extensions.LayoutContainer
 import kotlinx.android.synthetic.main.git_repository_card_view_holder.*
