+++ app/src/test/java/com/maximo/douglas/githubconsumer/components/gitrepositorycard/context/GitRepositoryCardTest.kt
@@ -6,7 +6,7 @@ import com.maximo.douglas.commons.utils.StringUtils.compactStringWithDots
 import com.maximo.douglas.domain.entity.gitrepository.GitRepository
 import com.maximo.douglas.domain.entity.user.GitUser
 import com.maximo.douglas.githubconsumer.testutils.BaseComponentTest
-import com.maximo.douglas.features.components.repositorycard.GitRepositoryCardComponent
+import com.maximo.douglas.githubconsumer.ui.components.repositorycard.GitRepositoryCardComponent
 import kotlinx.android.synthetic.main.git_repository_card_component.view.*
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertTrue
