+++ features/src/androidTest/java/com/maximo/douglas/features/InstrumentedTestSuite.kt
@@ -1,13 +1,13 @@
 package com.maximo.douglas.features
 
-import com.maximo.douglas.features.gitpullrequests.SelectGitPullRequestFragmentTest
-import com.maximo.douglas.features.gitrepositories.SelectGitRepositoryFragmentTest
+import com.maximo.douglas.githubconsumer.features.gitpullrequests.SelectGitPullRequestFragmentTest
+import com.maximo.douglas.githubconsumer.features.gitrepositories.SelectGitRepositoryFragmentTest
 import org.junit.runner.RunWith
 import org.junit.runners.Suite
 
 @RunWith(Suite::class)
 @Suite.SuiteClasses(
-    SelectGitRepositoryFragmentTest::class,
-    SelectGitPullRequestFragmentTest::class
+    com.maximo.douglas.githubconsumer.features.gitrepositories.SelectGitRepositoryFragmentTest::class,
+    com.maximo.douglas.githubconsumer.features.gitpullrequests.SelectGitPullRequestFragmentTest::class
 )
 class InstrumentedTestSuite
