--- features/src/main/res/drawable/ic_fork.xml
@@ -1,12 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="20dp"
-    android:height="20dp"
-    android:tint="@color/yellow700"
-    android:viewportWidth="24"
-    android:viewportHeight="24">
-
-    <path
-        android:fillColor="#FF000000"
-        android:pathData="M21,3c0,-1.657 -1.343,-3 -3,-3s-3,1.343 -3,3c0,1.323 0.861,2.433 2.05,2.832 0.168,4.295 -2.021,4.764 -4.998,5.391 -1.709,0.36 -3.642,0.775 -5.052,2.085v-7.492c1.163,-0.413 2,-1.511 2,-2.816 0,-1.657 -1.343,-3 -3,-3s-3,1.343 -3,3c0,1.305 0.837,2.403 2,2.816v12.367c-1.163,0.414 -2,1.512 -2,2.817 0,1.657 1.343,3 3,3s3,-1.343 3,-3c0,-1.295 -0.824,-2.388 -1.973,-2.808 0.27,-3.922 2.57,-4.408 5.437,-5.012 3.038,-0.64 6.774,-1.442 6.579,-7.377 1.141,-0.425 1.957,-1.514 1.957,-2.803zM4.2,3c0,-0.993 0.807,-1.8 1.8,-1.8s1.8,0.807 1.8,1.8 -0.807,1.8 -1.8,1.8 -1.8,-0.807 -1.8,-1.8zM7.8,21c0,0.993 -0.807,1.8 -1.8,1.8s-1.8,-0.807 -1.8,-1.8 0.807,-1.8 1.8,-1.8 1.8,0.807 1.8,1.8zM18,4.8c-0.993,0 -1.8,-0.807 -1.8,-1.8s0.807,-1.8 1.8,-1.8 1.8,0.807 1.8,1.8 -0.807,1.8 -1.8,1.8z" />
-
-</vector>
