+++ features/src/main/res/layout/git_pull_request_card_view_holder.xml
@@ -2,7 +2,7 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
-    <com.maximo.douglas.features.components.pullrequestcard.GitPullRequestCardComponent
+    <com.maximo.douglas.githubconsumer.ui.components.pullrequestcard.GitPullRequestCardComponent
         android:id="@+id/gitPullRequestCard"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
