+++ features/src/main/res/layout/git_repository_card_view_holder.xml
@@ -2,7 +2,7 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
-    <com.maximo.douglas.features.components.repositorycard.GitRepositoryCardComponent
+    <com.maximo.douglas.githubconsumer.ui.components.repositorycard.GitRepositoryCardComponent
         android:id="@+id/gitRepositoryCard"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
