+++ settings.gradle
@@ -1,4 +1,3 @@
-include ':features'
 include ':commons'
 include ':data'
 include ':domain'
